requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-storage
torch
torchvision
opencv-python-headless
python-dotenv